Refactor code structure for improved readability and maintainability

+++ Models/Transaction.cs
@@ -9,7 +9,7 @@ public class Transaction
     public int transaction_id { get; set; }
     public string description { get; set; } = string.Empty;
     public string transaction_type { get; set; } = string.Empty;
-    public float transaction_nominal { get; set; } = 0.0f;
+    public decimal transaction_nominal { get; set; } = 0;
     public DateTime transaction_date { get; set; } = DateTime.UtcNow;
     public int user_id { get; set; }
